// projects/laji/src/app/shared/model/LajiTaxonSearch.ts
/* eslint-disable max-len */
/**
 * API documentation
 * To use this api you need an access token. To getList the token, send a post request with your email address to api-users resource and one will be send to your. See below for information on how to use this api and if you have any questions you can contact us at helpdesk@laji.fi.  Place refer to [schema.laji.fi](http://schema.laji.fi/) for more information about the used vocabulary
 *
 * OpenAPI spec version: 0.0.1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

'use strict';

import { LajiInformalGroup } from './LajiInformalGroup';

export interface LajiTaxonSearch {


  /**
   * Unique identifier for the taxon
   */
  id?: string;

  /**
   * What kind of match this is
   */
    type?: string;

  /**
   * Name that the search matches to
   */
  matchingName?: string;

  /**
   * Scientific name for the match
   */
  scientificName?: string;

  /**
   * Author
   */
  scientificNameAuthorship?: string;

  /**
   * Taxon rank
   */
  taxonRankId?: string;

  /**
   * Score for the match
   */
  similarity?: string;

  /**
   * Informal group
   */
  informalGroup?: Array<LajiInformalGroup>;
}